remove saving logits unneccessarily for non-segmentation inferers (e.g. scribbles)

+++ sample-apps/segmentation_spleen_postproc/main.py
@@ -65,7 +65,7 @@ class MyApp(MONAILabelApp):
 
         # save logits
         logits = result_params.get("logits")
-        if logits:
+        if logits and self._infers[request.get("model")].type == InferType.SEGMENTATION:
             self.datastore().save_label(image, logits, "logits")
             os.unlink(logits)
